Fixed import errors and moved Apollo client to separate file
