--- src/modules/createEvent/eventTypes.ts
@@ -1,7 +1,0 @@
-export type Event =
-  | 'Hike'
-  | 'BikeRide'
-  | 'Canoeing'
-  | 'TroopMeeting'
-  | 'Campout'
-  | 'SummerCamp';
